requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
google-api-python-client>=2.70.0
youtube-transcript-api>=0.6.0
httpx>=0.24
python-dotenv>=1.0.0
langchain-openai>=0.0.0  # only if you will use the same LangChain LLM code; optional
openai>=0.27.0           # optional if using OpenAI directly
pandas>=2.0
python-dateutil
tqdm
langchain
langchain-community
# systemd-python>=235  # for systemd journal logging (optional, install on Linux)
requests>=2.31.0  # for HTTP requests to supadata.ai and other APIs